+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Revision history for cli-extras
 
+## 0.1.0.1
+
+* Loosen version bounds. Support GHC 8.8.4.
+* Remove orphan MonadFail instance for LoggingT (we now require monad-logger >= 0.3.30)
+
 ## 0.1.0.0
 * Initial release